Refactored client-websocket to its own package
